+++ crates/swc/tests/tsc-references/ModuleWithExportedAndNonExportedClasses_es5.1.normal.js
@@ -4,17 +4,17 @@ function _classCallCheck(instance, Constructor) {
     }
 }
 var A;
-(function(A3) {
-    var A1 = function A1() {
+(function(A1) {
+    var _$A = function _$A() {
         "use strict";
-        _classCallCheck(this, A1);
+        _classCallCheck(this, _$A);
     };
-    A3.A = A1;
+    A1.A = _$A;
     var AG1 = function AG1() {
         "use strict";
         _classCallCheck(this, AG1);
     };
-    A3.AG = AG1;
+    A1.AG = AG1;
     var A2 = function A2() {
         "use strict";
         _classCallCheck(this, A2);