+++ crates/swc/tests/tsc-references/asyncAwait_es6_es2015.1.normal.js
@@ -149,14 +149,14 @@ class C {
 }
 var M;
 (function(M1) {
-    function f11() {
+    function f15() {
         return _f11.apply(this, arguments);
     }
     function _f11() {
         _f11 = _asyncToGenerator(function*() {});
         return _f11.apply(this, arguments);
     }
-    M1.f1 = f11;
+    M1.f1 = f15;
 })(M || (M = {}));
 function f14() {
     return _f14.apply(this, arguments);