+++ crates/swc/tests/tsc-references/classExtendsItselfIndirectly2_es5.2.minified.js
@@ -76,7 +76,7 @@ var M, N, O, C = function(_E) {
     }(M.D);
     N1.E = E;
 })(N || (N = {})), (function(O) {
-    var P, Q2, P1, D2, Q1, E2, C2 = function(_E2) {
+    var P, _$Q, P1, D2, Q1, E2, C2 = function(_E2) {
         "use strict";
         _inherits(C2, _E2);
         var _super = _createSuper(C2);
@@ -93,7 +93,7 @@ var M, N, O, C = function(_E) {
             return _classCallCheck(this, D2), _super.apply(this, arguments);
         }
         return D2;
-    })(C2), P1.D2 = D2, Q1 = Q2 || (Q2 = {}), E2 = (function(_D2) {
+    })(C2), P1.D2 = D2, Q1 = _$Q || (_$Q = {}), E2 = (function(_D2) {
         "use strict";
         _inherits(E2, _D2);
         var _super = _createSuper(E2);