+++ crates/swc/tests/tsc-references/variadicTuples1_es5.1.normal.js
@@ -127,13 +127,13 @@ ft4([
 ]); // readonly [string, number]
 // Indexing variadic tuple types
 function f0(t, n) {
-    var a = t[0]; // string
+    var _$a = t[0]; // string
     var b = t[1]; // [string, ...T][1]
     var c = t[2]; // [string, ...T][2]
     var d = t[n]; // [string, ...T][number]
 }
 function f1(t, n) {
-    var a = t[0]; // string
+    var _$a = t[0]; // string
     var b = t[1]; // [string, ...T, number][1]
     var c = t[2]; // [string, ...T, number][2]
     var d = t[n]; // [string, ...T, number][number]
@@ -249,14 +249,14 @@ function f15(k0, k1, k2, k3) {
 }
 // Inference to [...T, ...U] with implied arity for T
 function curry(f) {
-    for(var _len1 = arguments.length, a = new Array(_len1 > 1 ? _len1 - 1 : 0), _key1 = 1; _key1 < _len1; _key1++){
-        a[_key1 - 1] = arguments[_key1];
+    for(var _len1 = arguments.length, _$a = new Array(_len1 > 1 ? _len1 - 1 : 0), _key1 = 1; _key1 < _len1; _key1++){
+        _$a[_key1 - 1] = arguments[_key1];
     }
     return function() {
-        for(var _len = arguments.length, b1 = new Array(_len), _key = 0; _key < _len; _key++){
-            b1[_key] = arguments[_key];
+        for(var _len = arguments.length, b2 = new Array(_len), _key = 0; _key < _len; _key++){
+            b2[_key] = arguments[_key];
         }
-        return f.apply(void 0, _toConsumableArray(a).concat(_toConsumableArray(b1)));
+        return f.apply(void 0, _toConsumableArray(_$a).concat(_toConsumableArray(b2)));
     };
 }
 var fn1 = function(a, b, c, d) {