+++ crates/swc_ecma_transforms_compat/tests/block-scoping/issue-3235/1/output.js
@@ -1,0 +1,16 @@
+var res = [];
+var a = 2;
+res.push(a === 2);
+{
+    var _$b = 1;
+}res.push(typeof b === "undefined");
+if (true) {
+    var _$b1 = 0;
+}
+res.push(typeof b === "undefined");
+for(var _$b2 = 0; _$b2 < 10; _$b2++){}
+res.push(typeof b === "undefined");
+function test() {
+    var _$b = 7;
+}
+res.push(typeof b === "undefined");