+++ crates/swc_ecma_transforms_compat/tests/es2015_classes.rs
@@ -2911,39 +2911,39 @@ class TestMultipleMethods extends (class {
 
 "#,
     r#"
-    var TestEmpty = function(_super) {
+    var TestEmpty = function(_superClass) {
       "use strict";
-      _inherits(TestEmpty, _super);
-      var _super1 = _createSuper(TestEmpty);
+      _inherits(TestEmpty, _superClass);
+      var _super = _createSuper(TestEmpty);
       function TestEmpty() {
           _classCallCheck(this, TestEmpty);
-          return _super1.apply(this, arguments);
+          return _super.apply(this, arguments);
       }
       return TestEmpty;
   }(function _class() {
       "use strict";
       _classCallCheck(this, _class);
   });
-  var TestConstructorOnly = function(_super) {
+  var TestConstructorOnly = function(_superClass) {
       "use strict";
-      _inherits(TestConstructorOnly, _super);
-      var _super2 = _createSuper(TestConstructorOnly);
+      _inherits(TestConstructorOnly, _superClass);
+      var _super = _createSuper(TestConstructorOnly);
       function TestConstructorOnly() {
           _classCallCheck(this, TestConstructorOnly);
-          return _super2.apply(this, arguments);
+          return _super.apply(this, arguments);
       }
       return TestConstructorOnly;
   }(function _class() {
       "use strict";
       _classCallCheck(this, _class);
   });
-  var TestMethodOnly = function(_super) {
+  var TestMethodOnly = function(_superClass) {
       "use strict";
-      _inherits(TestMethodOnly, _super);
-      var _super3 = _createSuper(TestMethodOnly);
+      _inherits(TestMethodOnly, _superClass);
+      var _super = _createSuper(TestMethodOnly);
       function TestMethodOnly() {
           _classCallCheck(this, TestMethodOnly);
-          return _super3.apply(this, arguments);
+          return _super.apply(this, arguments);
       }
       return TestMethodOnly;
   }(function() {
@@ -2960,13 +2960,13 @@ class TestMultipleMethods extends (class {
       ]);
       return _class;
   }());
-  var TestConstructorAndMethod = function(_super) {
+  var TestConstructorAndMethod = function(_superClass) {
       "use strict";
-      _inherits(TestConstructorAndMethod, _super);
-      var _super4 = _createSuper(TestConstructorAndMethod);
+      _inherits(TestConstructorAndMethod, _superClass);
+      var _super = _createSuper(TestConstructorAndMethod);
       function TestConstructorAndMethod() {
           _classCallCheck(this, TestConstructorAndMethod);
-          return _super4.apply(this, arguments);
+          return _super.apply(this, arguments);
       }
       return TestConstructorAndMethod;
   }(function() {
@@ -2983,13 +2983,13 @@ class TestMultipleMethods extends (class {
       ]);
       return _class;
   }());
-  var TestMultipleMethods = function(_super) {
+  var TestMultipleMethods = function(_superClass) {
       "use strict";
-      _inherits(TestMultipleMethods, _super);
-      var _super5 = _createSuper(TestMultipleMethods);
+      _inherits(TestMultipleMethods, _superClass);
+      var _super = _createSuper(TestMultipleMethods);
       function TestMultipleMethods() {
           _classCallCheck(this, TestMultipleMethods);
-          return _super5.apply(this, arguments);
+          return _super.apply(this, arguments);
       }
       return TestMultipleMethods;
   }(function() {
@@ -6761,7 +6761,7 @@ var Extended = function(Base) {
     "
 );
 
-#[testing::fixture("tests/fixture/classes/**/exec.js")]
+#[testing::fixture("tests/classes/**/exec.js")]
 fn exec(input: PathBuf) {
     let src = read_to_string(&input).unwrap();
     compare_stdout(
@@ -6776,7 +6776,7 @@ fn exec(input: PathBuf) {
     );
 }
 
-#[testing::fixture("tests/fixture/classes/**/input.js")]
+#[testing::fixture("tests/classes/**/input.js")]
 fn fixture(input: PathBuf) {
     let output = input.parent().unwrap().join("output.js");